smart_agriculture_api/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
scikit-learn==1.6.1
tensorflow==2.20.0
tensorboard==2.20.0
python-dotenv==1.0.1
pyserial==3.5